camera_service/requirements.txt
numpy==1.25.2
opencv_python==4.8.0.76
PyYAML==6.0.1
python-time
redis
multiprocess
jsons
